pyproject.toml
[project]
name = "pii-remover-analyser"
version = "0.1.0"
description = "A streamlit ui based project to analyse files with AI after removing any PII"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "python-dotenv>=0.9.9",
    "google-genai>=1.36.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "pip>=25.2",
    "presidio-analyzer>=2.2.360",
    "presidio-anonymizer>=2.2.360",
    "presidio-image-redactor>=0.0.57",
    "pypdf>=4.0.0",
    "pypdfium2>=4.0.0",
    "python-pptx>=1.0.2",
    "setuptools>=80.9.0",
    "spacy>=3.8.7",
    "st-annotated-text>=4.0.2",
    "streamlit>=1.49.1",
    "streamlit-tags>=1.2.8",
    "wheel>=0.45.1",
    "en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl",
    "pipreqs>=0.5.0",
]
